third_party/blink/web_tests/external/wpt/css/css-ui/parsing/outline-color-computed-expected.txt

This is a testharness.js-based test.
[FAIL] auto computes to itself or to currentColor depending on the value of outline-style
  assert_equals: expected "auto" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
